refactor: split up the projects section
