+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 import { AboutSection } from "@/components/sections/about/about-section";
 import { ContactSection } from "@/components/sections/contact-section";
 import { HomeSection } from "@/components/sections/home/home-section";
-import { ProjectsSection } from "@/components/sections/projects-section";
+import { ProjectsSection } from "@/components/sections/projects/projects-section";
 
 export default function Home() {
   return (
